pyproject.toml
[project]
name = "multilang"
version = "0.1.0"
description = "Aligns a text with its translation, semantically, on printable pages"
authors = [
    {name = "Lavinia Lee",email = "lavinia.lee@mail.mcgill.ca"}
]
readme = "README.md"
requires-python = ">=3.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
python-lorem = "^1.3.0.post3"


[tool.poetry.dependencies]
flask = "^3.1.2"
requests = "^2.32.5"
